♻️ refactor rspec files
I will implement shared_context, shared_examples in the future.
Before that, I would like to refactor rspec files to map library
files to rspec files.

This structure is supported by rspec-mode of emacs or "rails go to
spec" plugin of vscode.
